refactor: core:common util
