+++ feature/ocr/src/main/java/com/moneymong/moneymong/ocr/OCRScreen.kt
@@ -33,9 +33,9 @@ import com.moneymong.moneymong.design_system.component.modal.MDSModal
 import com.moneymong.moneymong.design_system.theme.Heading1
 import com.moneymong.moneymong.design_system.theme.Mint03
 import com.moneymong.moneymong.design_system.theme.White
-import com.moneymong.moneymong.common.ext.hasPermission
+import com.moneymong.moneymong.common.util.hasPermission
 import com.moneymong.moneymong.ui.noRippleClickable
-import com.moneymong.moneymong.common.util.DisposableEffectWithLifeCycle
+import com.moneymong.moneymong.ui.util.DisposableEffectWithLifeCycle
 import com.moneymong.moneymong.design_system.error.ErrorDialog
 import com.moneymong.moneymong.design_system.theme.Black
 import com.moneymong.moneymong.ocr.view.OCRCameraPermissionDeniedView
